refactor: enum 타입 패키지 변경 (#36)
